Refactor: #41 TextBox를 TitleInput, ContentsTextArea로 분리
